+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -15,7 +15,6 @@ import Destroyable from '../destroyable/Destroyable';
 import AudioVideoEventAttributes from '../eventcontroller/AudioVideoEventAttributes';
 import DefaultEventController from '../eventcontroller/DefaultEventController';
 import EventController from '../eventcontroller/EventController';
-import EventReporter from '../eventreporter/EventReporter';
 import Logger from '../logger/Logger';
 import MediaStreamBroker from '../mediastreambroker/MediaStreamBroker';
 import MeetingSessionConfiguration from '../meetingsession/MeetingSessionConfiguration';
@@ -141,7 +140,7 @@ export default class DefaultAudioVideoController
     webSocketAdapter: WebSocketAdapter,
     mediaStreamBroker: MediaStreamBroker,
     reconnectController: ReconnectController,
-    eventReporter?: EventReporter
+    eventController?: EventController
   ) {
     this._logger = logger;
     this.sessionStateController = new DefaultSessionStateController(this._logger);
@@ -166,7 +165,11 @@ export default class DefaultAudioVideoController
     );
     this._audioMixController = new DefaultAudioMixController(this._logger);
     this.meetingSessionContext.logger = this._logger;
-    this._eventController = new DefaultEventController(this, eventReporter);
+    if (eventController) {
+      this._eventController = eventController;
+    } else {
+      this._eventController = new DefaultEventController(configuration, logger);
+    }
   }
 
   async destroy(): Promise<void> {
@@ -202,10 +205,6 @@ export default class DefaultAudioVideoController
     return this._audioMixController;
   }
 
-  get eventController(): EventController {
-    return this._eventController;
-  }
-
   get logger(): Logger {
     return this._logger;
   }
@@ -216,6 +215,10 @@ export default class DefaultAudioVideoController
 
   get mediaStreamBroker(): MediaStreamBroker {
     return this._mediaStreamBroker;
+  }
+
+  get eventController(): EventController {
+    return this._eventController;
   }
 
   getRTCPeerConnectionStats(selector?: MediaStreamTrack): Promise<RTCStatsReport> {
@@ -597,10 +600,7 @@ export default class DefaultAudioVideoController
       this.forEachObserver(observer => {
         Maybe.of(observer.audioVideoDidStartConnecting).map(f => f.bind(observer)(false));
       });
-      /* istanbul ignore else */
-      if (this.eventController) {
-        this.eventController.publishEvent('meetingStartRequested');
-      }
+      this.eventController.publishEvent('meetingStartRequested');
     }
 
     this.meetingSessionContext.startAudioVideoTimestamp = this.startAudioVideoTimestamp;