+++ src/audiovideocontroller/NoOpAudioVideoController.ts
@@ -3,6 +3,7 @@
 
 import AudioProfile from '../audioprofile/AudioProfile';
 import FullJitterBackoff from '../backoff/FullJitterBackoff';
+import DefaultEventController from '../eventcontroller/DefaultEventController';
 import NoOpDebugLogger from '../logger/NoOpDebugLogger';
 import NoOpMediaStreamBroker from '../mediastreambroker/NoOpMediaStreamBroker';
 import MeetingSessionConfiguration from '../meetingsession/MeetingSessionConfiguration';
@@ -27,12 +28,14 @@ export default class NoOpAudioVideoController extends DefaultAudioVideoControlle
     emptyConfiguration.urls.screenDataURL = '';
     emptyConfiguration.urls.screenSharingURL = 'wss://localhost/';
     emptyConfiguration.urls.signalingURL = 'wss://localhost/';
+    const noOpLogger = new NoOpDebugLogger();
     super(
       configuration ? configuration : emptyConfiguration,
-      new NoOpDebugLogger(),
+      noOpLogger,
       new DefaultWebSocketAdapter(new NoOpDebugLogger()),
       new NoOpMediaStreamBroker(),
-      new DefaultReconnectController(0, new FullJitterBackoff(0, 0, 0))
+      new DefaultReconnectController(0, new FullJitterBackoff(0, 0, 0)),
+      new DefaultEventController(configuration ? configuration : emptyConfiguration, noOpLogger)
     );
   }